requirements.txt
attr==0.3.2
beautifulsoup4==4.12.3
config_file==0.13.1
ConfigParser==7.1.0
cryptography==44.0.0
dl==0.1.0
docutils==0.21.2
gensim==4.3.3
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.4
kafka_python==2.0.2
keras==3.7.0
keyring==25.5.0
lockfile==0.12.2
nltk==3.9.1
numpy==1.23.5
pandas==1.5.3
Pillow==11.0.0
protobuf==5.29.0
pyfiglet==1.0.2
pyOpenSSL==24.3.0
pyspark==3.5.3
python-dotenv==1.0.1
python-telegram-bot==21.7
railroad==0.5.0
Sphinx==8.1.3
Telethon==1.38.1
tensorflow
textblob==0.18.0.post0
thread==2.0.5
tornado==6.4.2
trove_classifiers==2024.10.21.16
truststore==0.10.0
tweepy==4.14.0
urllib3_secure_extra==0.1.0
vaderSentiment==3.3.2
wordcloud==1.9.4